otl-fastapi/requirements.txt
fastapi
uvicorn
opentelemetry-api
opentelemetry-sdk